# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API DOCUMENTATION",
        "title": "API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "prastiyo.beka12@gmail.com"
        }
    },
    "host": "localhost:3000",
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Login to system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth System"
                ],
                "summary": "Login System",
                "parameters": [
                    {
                        "description": "Login Input",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core.SignInInput"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/auth/sign-up": {
            "post": {
                "description": "Sign Up",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth System"
                ],
                "summary": "Sign Up User",
                "parameters": [
                    {
                        "description": "SignUpInput",
                        "name": "sign-up",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core.SignUpInput"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/company": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Tps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tps"
                ],
                "summary": "Tps",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Company Create",
                "parameters": [
                    {
                        "description": "Company Create",
                        "name": "company",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Company.CompanyRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/company/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Company Delete",
                "parameters": [
                    {
                        "description": "Company Update",
                        "name": "company",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Company.CompanyUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/company/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Company By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Company By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Company Update",
                "parameters": [
                    {
                        "description": "Company Update",
                        "name": "company",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Company.CompanyUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "Company.CompanyRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "dapil_id": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path_photo": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "Company.CompanyUpdateRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "dapil_id": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path_photo": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "core.SignInInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "core.SignUpInput": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "passwordConfirm",
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "passwordConfirm": {
                    "type": "string",
                    "minLength": 8
                },
                "phone": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}